Refactor: MonsterPath 부분 싱글톤 패턴으로 따로 분리
